trading-assistant/requirements.txt
requests
pandas
feedparser
textblob
streamlit
rich
yfinance<1.0
trafilatura
python-dotenv
readability-lxml
newspaper3k
beautifulsoup4
lxml_html_clean
